app/requirements.txt
aldryn-apphooks-config==0.6.0
aldryn-boilerplates==0.8.0
aldryn-common==1.0.5
amqp==1.4.9
anyjson==0.3.3
asgiref==3.6.0
billiard==3.3.0.23
calmjs.parse==1.3.0
celery-haystack==0.10
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
Django==4.2.9
django-appconf==1.0.5
django-appdata==0.3.2
django-classy-tags==4.0.0
django-cms==3.11.4
django-colorfield==0.8.0
django-compressor==4.3.1
django-cors-headers==4.3.0
django-environ==0.10.0
django-filer==3.1.0
django-filter==23.1
django-formtools==2.4
django-fsm==2.8.1
django-haystack==3.2.1
django-js-asset==2.0.0
django-meta==2.1.0
django-mptt==0.14.0
django-parler==2.3
django-phonenumber-field==7.1.0
django-polymorphic==3.1.0
django-prometheus==2.3.1
django-recaptcha==3.0.0
django-sekizai==4.0.0
django-sortedm2m==3.1.1
django-spurl==0.6.8
django-standard-form==1.1.1
django-taggit==3.1.0
django-taggit-autosuggest==0.4.1
django-taggit-labels==0.7.0
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-theherk-menu-icons==2.0.1
django-treebeard==4.4
djangocms-4-migration==0.0.2
djangocms-admin-style==3.2.3
djangocms-apphook-setup==0.4.1
djangocms-attributes-field==2.1.0
djangocms-blog==1.2.3
djangocms-link==3.1.0
djangocms-picture==4.0.0
djangocms-text-ckeditor==5.1.5
djangocms-versioning==2.0.0
djangocms-video==3.0.0
easy-thumbnails==2.8.5
elastic-transport==8.4.0
elasticsearch==7.17.9
gunicorn==21.2.0
html5lib==1.1
idna==3.4
kombu==3.0.37
lxml==4.9.2
numpy==1.25.2
packaging==23.0
phonenumberslite==8.13.18
Pillow==9.4.0
ply==3.11
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg==3.1.9
psycopg-binary==3.1.9
pysolr==3.9.0
python-dateutil==2.8.2
pytz==2022.7.1
rcssmin==1.1.1
reportlab==4.0.4
requests==2.31.0
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
typing-extensions==4.7.1
tzdata==2023.3
Unidecode==1.1.2
urllib3==1.26.16
URLObject==2.4.3
vine==1.3.0
wcwidth==0.2.6
webencodings==0.5.1
YURL==1.0.0
